requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.22.0
PyYAML==6.0.2
numpy==1.26.4
streamlit-authenticator==0.3.2
scikit-learn==1.5.1
requests==2.32.3
bcrypt==4.1.3
fpdf2==2.7.9
